% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-exhibits-core.R
\name{descriptive_expand_category}
\alias{descriptive_expand_category}
\title{Expand a Categorical Column into Indicator Dummies}
\usage{
descriptive_expand_category(data, category_var)
}
\arguments{
\item{data}{\code{data.frame}.}

\item{category_var}{Character. Categorical column, e.g. \code{"LndOwn"}.}
}
\value{
\code{data} with the dummy columns appended. The new names are returned in
\code{attr(x, "indicators")}.
}
\description{
Reproduces Stata's \verb{tab <var>, gen(<var>_)}: one 0/1 column per level, named
\verb{<var>_1}, \verb{<var>_2}, ... in level order.
}
\seealso{
Other descriptive exhibits:
\code{\link[=descriptive_group_summary]{descriptive_group_summary()}},
\code{\link[=descriptive_indicator_shares]{descriptive_indicator_shares()}},
\code{\link[=descriptive_prepare]{descriptive_prepare()}},
\code{\link[=descriptive_season_year]{descriptive_season_year()}},
\code{\link[=descriptive_specifications]{descriptive_specifications()}},
\code{\link[=descriptive_trend_model]{descriptive_trend_model()}},
\code{\link[=descriptive_workhorse]{descriptive_workhorse()}},
\code{\link[=draw_descriptive_summary]{draw_descriptive_summary()}}
}
\concept{descriptive exhibits}
