% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{harmonized_data_prep}
\alias{harmonized_data_prep}
\title{Prepare Data for Agricultural Productivity Analysis}
\usage{
harmonized_data_prep(data)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} containing household-
and farm-level variables such as weights, demographic information,
and agricultural inputs.}
}
\value{
A cleaned and transformed \code{data.frame} or \code{data.table}
with additional variables ready for analysis.
}
\description{
Cleans and transforms a dataset by creating new variables, applying
log transformations, converting selected variables to factors or
characters, and recoding education levels. The function is designed
to standardize inputs for further analysis of agricultural productivity.
}
\seealso{
Other helpers: 
\code{\link{compute_jenks_binned_shares}()},
\code{\link{get_crop_area_list}()}
}
\concept{helpers}
